Transitions working
Got basic transition working - I decided to go with the
react-transition-group library in the end as the M-ui transitions were
less configurable.

Also downgraded M-ui to V4 - M-ui v5 came out in the last couple of days
and It looks like it's got loads of bugs! I tried to get it working for
a while and then decided to go with v4 for this project.

Set browser tab icon