Featured Destinations Finished
And fully responsive down to 300px

I had to move the value hook for the header selection out of the header
component so that the 'view All' link in <FeaturedDestinations /> can
access it. It is now drilled through from App.js.

I also noticed that the selection underline was being rendered before
the <Animate /> transition had finished and as a result was being drawn
in the wrong place. I have fixed this by introducing a delay before it
renders, it also looks cool!