All buttons link to About
I have had to make a ScrollToTop component so when changing route you
can reset the page scroll position.

I have also hade to make the header appear only when visible, it looked
wrong just appearing there when you change the route.

This also brought back the problem with the wonky underline bar! I fixed
it by moving the setTimeout() in to the setHeaderUnderline() function
and removing it from useEffect. This is a more elegant solution to the
problem anyway. There is now a second input to the setHeaderUnderline()
function that defaults to false but if true will impose a delay on
re-positioning the header underline.

I have left the dependancy array for useEffect blank because it must
only be run once I.e when you navigate to the page from a URL. It throws
a linter warning but I can't find anything to say that it is bad.