Review slider now working
It all went together pretty seamlessly in the end, there is one thing
left to do though which is the cross and circle in the top and bottom
corners need placing in the correct position. I had a quick go but the
Card element is not the full size of the root component so when the
window resizes it all moves around.

I have hidden the decal component for now, it's late, off to bed!